# Swagger/swagg.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "MovieTicketBooking(MTB) Application",
        "description": "API document for MovieTicketBooking Service",
        "x-task1-appname": "MTB",
        "x-task1-seal": 104437,
        "version": "23.03.0"
    },
...
    "servers": [
        {
            "url": "",
            "description": "http://localhost:1111-Relative server url",
            "variables": {}
        }
    ],
    "paths": {
...
        "/admin": {
            "get": {
                "tags": [
                    "ADMIN TABLE"
                ],
                "summary": "Get all admins",
                "operationId": "getAdmins",
                "responses": {
                    "200": {
                        "description": "Returns a list of admins",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Admin"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Auth Token",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Request is not valid",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "bbt": [],
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ADMIN TABLE"
                ],
                "summary": "Create a new admin",
                "operationId": "createAdmin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Admin"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin created successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Admin"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request is not valid",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Auth Token",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "bbt": [],
                        "token": []
                    }
                ]
            }
        },
        "/admin/{AdminID}": {
            "get": {
                "tags": [
                    "ADMIN TABLE"
                ],
                "summary": "Get admin by AdminID",
                "operationId": "getAdminById",
                "parameters": [
                    {
                        "name": "AdminID",
                        "in": "path",
                        "description": "Admin ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the admin details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Admin"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request is not valid",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Auth Token",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "bbt": [],
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ADMIN TABLE"
                ],
                "summary": "Update admin details",
                "operationId": "updateAdmin",
                "parameters": [
                    {
                        "name": "AdminID",
                        "in": "path",
                        "description": "Admin ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Admin"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin details updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Admin"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request is not valid",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Auth Token",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "bbt": [],
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ADMIN TABLE"
                ],
                "summary": "Delete admin",
                "operationId": "deleteAdmin",
                "parameters": [
                    {
                        "name": "AdminID",
                        "in": "path",
                        "description": "Admin ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Admin deleted successfully"
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Auth Token",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Request is not valid",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "bbt": [],
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Admin": {
                "type": "object",
                "properties": {
                    "AdminID": {
                        "type": "string"
                    },
                    "Name": {
                        "type": "string"
                    },
                    "Password": {
                        "type": "string"
                    },
                    "Email": {
                        "type": "string",
                        "format": "email"
                    },
                    "Phone": {
                        "type": "string"
                    }
                }
            },
...
            "ErrorInfo": {
                "type": "object",
                "properties": {
                    "errorCode": {
                        "type": "string"
                    },
                    "errorMessage": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "bbt": {
                "type": "apiKey",
                "in": "header",
                "name": "BBT"
            },
            "token": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization"
            }
        }
    }
}